pyproject.toml
[project]
name = "whisper-typing"
version = "0.0.0"
description = "An example Python project for generating a Fibonacci sequence"
readme = "README.md"
requires-python = ">=3.13"
license = "Unlicense"
authors = [
  { name = "Leonardo Bonacci", email = "leonardo@bonacci.com" },
]
keywords = ["example", "fibonacci"]
classifiers = [
  "Development Status :: 1 - Planning",
  "Programming Language :: Python :: 3 :: Only",
]
dependencies = [
    "accelerate>=1.12.0",
    "colorama>=0.4.6",
    "google-generativeai>=0.8.6",
    "numpy>=2.4.1",
    "optimum>=2.1.0",
    "pygetwindow>=0.0.9",
    "pynput>=1.8.1",
    "python-dotenv>=1.2.1",
    "scipy>=1.17.0",
    "sounddevice>=0.5.3",
    "torch>=2.9.1",
    "transformers>=4.57.6",
]

[project.urls]
issues = "https://github.com/leonardo/bonacci/issues"
repository = "https://github.com/leonardo/bonacci.git"

[project.scripts]
whisper-typing = "whisper_typing.__main__:main"

[dependency-groups]
dev = [
  "lefthook>=1.12.4",
  "pytest>=8.4.2",
  "pytest-cov>=7.0.0",
  "ruff>=0.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.coverage.report]
fail_under = 100
omit = ["__main__.py"]

[tool.coverage.run]
source = ["whisper_typing"]

[tool.hatch.build.targets.sdist]
include = ["src"]

[tool.ruff.lint]
select = ["ALL"]
ignore = ["COM812", "D203", "D213"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "D"]
